userspace: split manager.go into per-domain files (#4421)
pkg/dataplane/userspace/manager.go had grown to 1823 lines mixing
several unrelated concerns behind the Manager type: config
compilation and the policy-scheduler/snapshot-protocol gate, route
and feed overlay publishing, generation/FIB counters, neighbor
snapshot maintenance, and helper status/control operations. The HA
domain was already carved out into manager_ha.go; this continues that
decomposition for the remaining cross-domain grab-bag.

This is pure mechanical code motion. Whole declarations (methods keep
their receiver, plus their domain-local helper functions and types)
are moved verbatim into sibling files in the same package, so every
caller is unchanged:

  manager_compile.go     Compile, UpdatePolicyScheduleState, the
                         policy-scheduler active-state helpers, the
                         snapshot-protocol gate sentinels /
                         ensure*ProtocolLocked / disarm helpers,
                         syncInterfaceAttachments,
                         configHasScheduledPolicy
  manager_overlay.go     route/feed overlay snapshot + clone helpers
                         and PublishRouteOverlaySnapshot
  manager_generation.go  readFIBGeneration, bpfKtimeNs,
                         bumpGeneration, BumpFIBGeneration
  manager_neighbor.go    neighborIndexKey and the SnapshotNeighbor /
                         monitored-ifindex maintenance methods
  manager_status.go      recordHelperStatusLocked plus the
                         Status/SetForwardingArmed/SetQueueState/
                         SetBindingState/InjectPacket/DrainSessionDeltas/
                         ExportOwnerRGSessions control operations

The residual manager.go (421 lines) keeps the Manager struct, the
DataplaneMode enum, lifecycle/accessor methods (New/Boot/Start/Load/
Close/Teardown, Link/HA/Sessions/...), ApplyConfig, and
shouldAttemptRSTSuppression. Per-file import blocks were regenerated
with goimports; no declaration body changed.

shim_loader_boundary_test.go now reads Compile from
manager_compile.go (it scans a named source file for the function),
following the moved declaration.

Validation:
  - decltool AST extraction: 82 production top-level declarations in
    both origin/master manager.go and the union of the six split
    files; sorted decl-name sets diff empty.
  - per-declaration sha256 body multiset identical between pristine
    and split (byte-for-byte, imports excluded).
  - gofmt -l clean; go vet clean.
  - go build ./... green; go test ./pkg/dataplane/userspace/... green.
