+++ pkg/dataplane/userspace/manager.go
@@ -2,15 +2,8 @@ package userspace
 
 import (
 	"context"
-	"errors"
-	"fmt"
-	"log/slog"
-	"net"
-	"os"
 	"os/exec"
-	"path/filepath"
 	"slices"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -20,68 +13,10 @@ import (
 	"github.com/psaab/xpf/pkg/config"
 	"github.com/psaab/xpf/pkg/dataplane"
 	dpruntime "github.com/psaab/xpf/pkg/dataplane/runtime"
-	"github.com/vishvananda/netlink"
-	"golang.org/x/sys/unix"
 )
 
 var _ dataplane.ConfigSink = (*Manager)(nil)
 var _ dataplane.RuntimeDataPlane = (*Manager)(nil)
-
-var ErrPolicySchedulerProtocolIncompatible = errors.New("userspace policy scheduler snapshot protocol incompatible")
-var ErrPersistentSourceNATProtocolIncompatible = errors.New("userspace persistent source NAT snapshot protocol incompatible")
-
-// requiredProtocolGateSentinels enumerates every "this config cannot be
-// committed against the helper's current ConfigSnapshotProtocolVersion"
-// sentinel produced by ensureRequiredSnapshotProtocolLocked. ApplyConfig
-// disarms the helper (Armed=false, fail-closed) and returns one of these
-// when the running helper is too old to honor the committed config. A
-// commit that hits any of them MUST abort — i.e. the daemon must surface a
-// failed commit to the operator rather than report success against a
-// disarmed dataplane (#2138).
-//
-// The lenient-load doctrine (#1960) is unaffected, because abort changes
-// behavior ONLY for daemon callers that surface the apply error to a human:
-//   - Boot/restart of an already-persisted config goes through the void
-//     applyConfig wrapper, which logs slog.Warn and swallows the error —
-//     the node boots through (warn, not brick).
-//   - Peer config-sync goes through syncAndApply, which DOES propagate the
-//     error; its caller (handleConfigSync) logs slog.Error and returns the
-//     error to configApplyLoop, which counts ConfigsApplyFailed and leaves the
-//     config high-water mark UNADVANCED so the primary's re-push re-converges
-//     the standby (M-2/#4151). When SyncApply already promoted the store, the
-//     re-push short-circuits on the "already matches active" check and heals
-//     the high-water; the node stays consistent with the peer (helper
-//     disarmed, not bricked).
-//
-// Only the operator-facing commit path (commitAndApply /
-// commitConfirmedAndApply) returns the abort to the committer.
-//
-// Every future ensureRequiredSnapshotProtocolLocked gate MUST add its
-// sentinel here so the commit-abort policy can never silently omit it
-// (the omission this list exists to prevent was exactly #2138: the
-// persistent-source-NAT gate disarmed the helper but was missing from the
-// daemon's abort set).
-var requiredProtocolGateSentinels = []error{
-	ErrPolicySchedulerProtocolIncompatible,
-	ErrPersistentSourceNATProtocolIncompatible,
-}
-
-// IsRequiredProtocolGateError reports whether err is (or wraps) any
-// required helper-protocol gate sentinel — the set that must abort a
-// commit. The daemon commit policy (compileErrorMustAbortApply) delegates
-// to this so the abort set has a single source of truth co-located with
-// the sentinels and the ensureRequiredSnapshotProtocolLocked gate that
-// emits them.
-func IsRequiredProtocolGateError(err error) bool {
-	for _, sentinel := range requiredProtocolGateSentinels {
-		if errors.Is(err, sentinel) {
-			return true
-		}
-	}
-	return false
-}
-
-const persistentSourceNATHAUnsupportedReason = "userspace persistent-nat source pool leases are not HA-synchronized"
 
 // DataplaneMode describes which packet-processing pipeline is active.
 type DataplaneMode int
@@ -404,95 +339,6 @@ func (m *Manager) recordApplyResultLocked(result *dataplane.ApplyResult, caps Us
 	m.lastApply = result.Clone()
 }
 
-// recordPolicyContentRejectionLocked tracks the #3261 diagnostic for the
-// just-built snapshot: the reasons (if any) it carries unrepresentable policy
-// content that the helper integrity preflight will reject. It is called at the
-// snapshot-build site BEFORE the publish so it is recorded even when the
-// publish is rejected (the helper keeps previous-good / default-deny while
-// staying armed — never fail-open). A one-shot slog line fires only on a
-// transition (per the logging rules — NOT per apply): a Warn when content
-// becomes unrepresentable (the operator must see the snapshot is being
-// rejected) and an Info when it becomes representable again.
-func (m *Manager) recordPolicyContentRejectionLocked(reasons []string) {
-	had := len(m.lastSnapshotRejectReasons) > 0
-	now := len(reasons) > 0
-	m.lastSnapshotRejectReasons = append([]string(nil), reasons...)
-	switch {
-	case now && !had:
-		slog.Warn(
-			"userspace: snapshot carries unrepresentable policy content; the helper integrity preflight rejects it and retains the previous-good state (fresh boot: default-deny). Helper stays armed — no kernel fail-open. Edit out the offending application/address and re-commit to restore enforcement.",
-			"reasons", reasons,
-		)
-	case had && !now:
-		slog.Info("userspace: policy content is representable again; snapshot will publish normally")
-	}
-}
-
-// recordZoneIDCollisionsLocked stores the #3719 zone-id-collision diagnostic
-// from the last snapshot build and fires a one-shot operator alarm on a
-// transition (per the logging rules — NOT per apply). A collision reaches this
-// only on the LENIENT path (a tolerant load, an HA sync from an un-upgraded
-// peer, or a config a pre-#3075 binary persisted); the strict commit path
-// rejects it. The builder already QUARANTINED the later-sorting colliding zone
-// (dropped from the wire, its interfaces unzoned, its policies removed), so the
-// dataplane is fail-closed and never merges two zones — but zone isolation is
-// DEGRADED (the quarantined zone forwards nothing) until an operator renames
-// one zone, so this is a loud Error naming both zones.
-func (m *Manager) recordZoneIDCollisionsLocked(collisions []ZoneIDCollision) {
-	had := len(m.lastZoneIDCollisions) > 0
-	msgs := make([]string, 0, len(collisions))
-	for _, c := range collisions {
-		msgs = append(msgs, c.String())
-	}
-	now := len(msgs) > 0
-	m.lastZoneIDCollisions = msgs
-	switch {
-	case now && !had:
-		slog.Error(
-			"userspace: security-zone id collision — two zone names fold to the same StableZoneID; the later-sorting zone is QUARANTINED (dropped from the dataplane, its interfaces unzoned and its traffic denied) so two zones never share an id. Zone isolation is DEGRADED until one zone is renamed and the config re-committed.",
-			"collisions", msgs,
-		)
-	case had && !now:
-		slog.Info("userspace: security-zone id collision cleared; all zones install with distinct ids")
-	}
-}
-
-func copyPolicySchedulerActiveState(activeState map[string]bool) map[string]bool {
-	if activeState == nil {
-		return nil
-	}
-	out := make(map[string]bool, len(activeState))
-	for name, active := range activeState {
-		out[name] = active
-	}
-	return out
-}
-
-func (m *Manager) policySchedulerActiveStateSnapshot() map[string]bool {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	return copyPolicySchedulerActiveState(m.policySchedulerActive)
-}
-
-// PolicySchedulerActiveState returns a copy of the daemon-maintained
-// per-scheduler active-state map (scheduler name -> currently active).
-// Read-only show surfaces (#3062 CLI/gRPC policy detail) consult it via
-// PolicyInactive to render runtime scheduler-driven policy state without
-// recomputing wall-clock schedule windows. A nil result means no
-// scheduler state has been published yet.
-func (m *Manager) PolicySchedulerActiveState() map[string]bool {
-	return m.policySchedulerActiveStateSnapshot()
-}
-
-// SetPolicySchedulerActiveState seeds the active-state map used by the next
-// full snapshot build. The daemon calls this while holding applySem so config
-// commits and scheduler flips cannot publish hybrid policy snapshots.
-func (m *Manager) SetPolicySchedulerActiveState(activeState map[string]bool) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	m.policySchedulerActive = copyPolicySchedulerActiveState(activeState)
-}
-
 // EventStream returns the event stream instance, or nil if not available.
 func (m *Manager) EventStream() *EventStream {
 	m.mu.Lock()
@@ -572,1252 +418,4 @@ func (m *Manager) SetDeferWorkers(v bool) {
 	m.mu.Lock()
 	m.deferWorkers = v
 	m.mu.Unlock()
-}
-
-func (m *Manager) Compile(cfg *config.Config) (*dataplane.CompileResult, error) {
-	// Delete XDP link pins BEFORE CompileUserspaceShim() so AttachXDP does
-	// a fresh attach. This is critical for zero-copy: fresh attach
-	// triggers mlx5 to initialize XSK buffer pool from fill ring.
-	// Pinned link reuse (l.Update) only swaps the program without
-	// reinitializing XSK RQs, leaving the fill ring unconsumed.
-	if linkPinDir := "/sys/fs/bpf/xpf/links"; true {
-		entries, _ := os.ReadDir(linkPinDir)
-		for _, e := range entries {
-			if strings.HasPrefix(e.Name(), "xdp_") {
-				path := filepath.Join(linkPinDir, e.Name())
-				_ = os.Remove(path)
-			}
-		}
-	}
-	caps := deriveUserspaceCapabilities(cfg)
-	_ = caps // used below for helper config
-	// Userspace mode always attaches the retained XDP shim. The shim
-	// redirects to XSK when ctrl=1; when ctrl=0 it only passes proven
-	// local/control traffic to the kernel and drops transit. Do not swap to
-	// xdp_main_prog for unsupported capabilities or failed XSK liveness: the
-	// userspace runtime must not require the legacy main XDP pipeline.
-	m.bpfShim.SelectUserspaceXDPShimEntryProgram()
-	result, err := m.bpfShim.CompileUserspaceShim(cfg)
-	if err != nil {
-		return nil, err
-	}
-	ucfg := deriveUserspaceConfig(cfg)
-	activeState := m.policySchedulerActiveStateSnapshot()
-	// #1827: include the cached ip-monitoring route overlay so a full
-	// apply (operator commit) while a policy is FAILED preserves the
-	// injected route instead of reverting traffic to the dead uplink.
-	// #2514: a config-shaped input (e.g. address-book content-ID
-	// collision) must reject the apply with an error rather than panic
-	// the daemon. buildSnapshot* returns the error up here; ApplyConfig
-	// fails closed and the previously published snapshot / dataplane state
-	// is retained (m.lastSnapshot is not advanced on the error path).
-	snap, err := buildSnapshotWithSchedulerStateAndNATCounters(cfg, ucfg, m.bumpGeneration(), m.readFIBGeneration(), activeState, m.routeOverlaySnapshot(), m.feedSnapshotOverlay(), result.NATCounterIDs)
-	if err != nil {
-		return nil, fmt.Errorf("userspace: build config snapshot: %w", err)
-	}
-	// #1620: stamp the cold-path sample mask onto the snapshot. The
-	// daemon called SetColdPathSampleMask once at startup with the
-	// validated CLI flag value (or nil for "use default"). A nil
-	// pointer here leaves the wire field absent (omitempty), which
-	// the Rust receiver unwrap_or-s to 0xff per plan §4.3.
-	m.mu.Lock()
-	snap.ColdPathSampleMask = m.coldPathSampleMask
-	m.mu.Unlock()
-	m.syncInterfaceAttachments(result, snap)
-
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	// #3261: record whether this snapshot carries unrepresentable policy
-	// content BEFORE the publish, so the diagnostic is captured even when the
-	// helper rejects the snapshot (the publish path returns early on the
-	// integrity error, before recordApplyResultLocked). The helper stays armed
-	// for this class; the reject retains previous-good (or leaves the fresh-boot
-	// default-deny), never fail-open.
-	m.recordPolicyContentRejectionLocked(snap.Capabilities.PolicyContentRejected)
-	// #3719: record + alarm any StableZoneID collision the builder quarantined
-	// (lenient / HA-sync / pre-#3075-persisted path). The colliding zone was
-	// already dropped from snap; this surfaces the degraded-isolation state.
-	m.recordZoneIDCollisionsLocked(snap.zoneIDCollisions)
-	m.clusterHA = cfg != nil && cfg.Chassis.Cluster != nil
-	m.seedHAGroupInventoryLocked(cfg)
-	prevPlanKey := snapshotBindingPlanKey(m.lastSnapshot)
-	newPlanKey := snapshotBindingPlanKey(snap)
-	pendingXSKStartup := m.proc != nil &&
-		m.proc.Process != nil &&
-		m.publishedSnapshot != 0 &&
-		!m.xskLivenessProven &&
-		!m.xskLivenessFailed
-	samePlanRefresh := m.proc != nil &&
-		m.proc.Process != nil &&
-		prevPlanKey != "" &&
-		prevPlanKey == newPlanKey
-	publishedPlanChangedDuringStartup := pendingXSKStartup &&
-		m.publishedPlanKey != "" &&
-		m.publishedPlanKey != newPlanKey
-	if publishedPlanChangedDuringStartup {
-		slog.Info(
-			"userspace: restarting helper during XSK startup for binding plan change",
-			"generation", snap.Generation,
-			"fib_generation", snap.FIBGeneration,
-		)
-		m.stopLocked()
-		pendingXSKStartup = false
-		samePlanRefresh = false
-	}
-	// #1197 v4 (Codex code-review v3 #1+#2): rebuild listener
-	// caches ONLY after a successful apply_snapshot. Doing it
-	// here (before publish) leaves the listener thinking
-	// userspace-dp has entries it doesn't if apply_snapshot fails.
-	// Moved to the post-success path below (after line 343).
-	if pendingXSKStartup {
-		if err := m.ensureRequiredSnapshotProtocolLocked(cfg); err != nil {
-			if disarmErr := m.disarmSnapshotProtocolFailureLocked(err); disarmErr != nil {
-				return result, errors.Join(err, disarmErr)
-			}
-			return result, err
-		}
-		if err := m.syncUserspaceClassifierMapsFailClosedLocked(snap); err != nil {
-			return result, err
-		}
-		// #1928 (Codex review Q1): the deferred-publish resume path
-		// (syncSnapshotLocked in process.go) never syncs HA state, so a
-		// cluster->standalone reconfig that lands during the XSK-startup
-		// deferral window would leave stale HA groups in the helper and
-		// re-arm the HAInactive transit-drop gate. seedHAGroupInventoryLocked
-		// already cleared m.haGroups for the non-cluster case above; clear the
-		// helper side here too so the standalone state is consistent
-		// regardless of which apply path runs. (Idempotent empty update.)
-		if !m.clusterHA {
-			if err := m.clearHelperHAStateLocked(); err != nil {
-				return result, fmt.Errorf("clear userspace HA state (deferred startup): %w", err)
-			}
-		}
-		m.lastSnapshot = snap
-		m.cfg = ucfg
-		m.recordApplyResultLocked(dataplane.ApplyResultFromCompileResult(result), caps, snap.Generation)
-		slog.Info(
-			"userspace: deferring snapshot publish during XSK startup",
-			"generation", snap.Generation,
-			"fib_generation", snap.FIBGeneration,
-			"same_plan", samePlanRefresh,
-		)
-		return result, nil
-	}
-	if samePlanRefresh {
-		if err := m.syncUserspaceClassifierMapsFailClosedLocked(snap); err != nil {
-			return result, err
-		}
-	} else {
-		if err := m.programBootstrapMapsLocked(snap, ucfg); err != nil {
-			return result, err
-		}
-	}
-	if err := m.ensureProcessLocked(ucfg); err != nil {
-		return result, err
-	}
-	if err := m.ensureRequiredSnapshotProtocolLocked(cfg); err != nil {
-		if disarmErr := m.disarmSnapshotProtocolFailureLocked(err); disarmErr != nil {
-			return result, errors.Join(err, disarmErr)
-		}
-		return result, err
-	}
-	if m.deferWorkers {
-		snap.DeferWorkers = true
-	}
-	var status ProcessStatus
-	if err := m.disarmBeforeUnsupportedPublishLocked(snap); err != nil {
-		return result, err
-	}
-	// #1197 v5 (Codex code-review v4 #2): apply_snapshot must
-	// send publishable-only neighbors to match the
-	// update_neighbors path. Otherwise Rust's full-snapshot
-	// build accepts state="none" entries Go's predicate rejects,
-	// and Go can't track removal of those entries via the index.
-	publishSnap := *snap
-	publishSnap.Neighbors = filterPublishableNeighbors(snap.Neighbors)
-	if err := m.requestLocked(ControlRequest{Type: "apply_snapshot", Snapshot: &publishSnap}, &status); err != nil {
-		return result, fmt.Errorf("publish userspace snapshot: %w", err)
-	}
-	m.logWgEndpointSetTransitionLocked(&publishSnap, "apply")
-	m.lastSnapshot = snap
-	// #1197 v4: apply_snapshot succeeded — userspace-dp has the
-	// new neighbors. NOW rebuild listener caches; before this
-	// point the index would shadow events for entries the
-	// dataplane hadn't accepted.
-	m.rebuildNeighborIndex()
-	m.rebuildMonitoredIfindexes()
-	m.publishedSnapshot = snap.Generation
-	m.publishedPlanKey = newPlanKey
-	// #2079: this full apply_snapshot succeeded — record the applied
-	// (config, generation) for the NAT pool-utilization-alarm monitor.
-	m.markAppliedSnapshotLocked()
-	if h, ok := snapshotContentHash(snap); ok {
-		m.lastSnapshotHash = h
-	}
-	if err := m.applyHelperStatusLocked(&status); err != nil {
-		return result, fmt.Errorf("sync helper status: %w", err)
-	}
-	// #1928: HA group state must only be replayed/published for chassis-cluster
-	// members. The rg_active map is a fixed-size ARRAY (16 entries, keys 0-15)
-	// so it is ALWAYS fully populated — even on a standalone firewall with no
-	// redundancy groups, where every entry is inactive. On standalone the old
-	// unconditional refreshHAStateFromMapsLocked() therefore fabricated 16
-	// inactive HA groups and shipped them to the helper; the helper's per-packet
-	// HA gate (enforce_ha_resolution_snapshot) then treats every transit
-	// ForwardCandidate as HAInactive (owner_rg_id<=0 && !ha_state.is_empty())
-	// and drops it — a total transit forwarding outage on non-cluster nodes. The
-	// periodic status poll already guards the refresh with m.clusterHA (see
-	// process.go); the startup path must match.
-	if m.clusterHA {
-		if err := m.refreshHAStateFromMapsLocked(); err != nil {
-			return result, fmt.Errorf("replay userspace HA state from maps: %w", err)
-		}
-		if err := m.syncHAStateLocked(); err != nil {
-			return result, fmt.Errorf("publish userspace HA state: %w", err)
-		}
-	} else if err := m.clearHelperHAStateLocked(); err != nil {
-		// Non-cluster node: ensure neither the manager nor the helper retains
-		// HA groups. seedHAGroupInventoryLocked already cleared m.haGroups
-		// above; this also clears any groups a prior clustered apply pushed to
-		// the helper (cluster->standalone live reconfig), which would otherwise
-		// keep the HAInactive transit-drop gate armed (Codex review #1928 Q3).
-		return result, fmt.Errorf("clear userspace HA state: %w", err)
-	}
-	if err := m.syncDesiredForwardingStateLocked(); err != nil {
-		return result, fmt.Errorf("sync userspace forwarding state: %w", err)
-	}
-	m.ensureStatusLoopLocked()
-	m.cfg = ucfg
-	m.recordApplyResultLocked(dataplane.ApplyResultFromCompileResult(result), caps, snap.Generation)
-	return result, nil
-}
-
-// UpdatePolicyScheduleState republishes the userspace policy snapshot with one
-// coherent inactive-bit view. This shadows the embedded eBPF manager method;
-// scheduled userspace policies must not update the policy_rules BPF map.
-func (m *Manager) UpdatePolicyScheduleState(cfg *config.Config, activeState map[string]bool) error {
-	activeCopy := copyPolicySchedulerActiveState(activeState)
-
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	m.policySchedulerActive = activeCopy
-	if cfg == nil {
-		if m.lastSnapshot == nil {
-			// #3780: no snapshot ever published — nothing to
-			// republish, and no live enforcement to go stale. The
-			// next full apply publishes the initial state. Converged.
-			return nil
-		}
-		cfg = m.lastSnapshot.Config
-	}
-	if cfg == nil || m.lastSnapshot == nil {
-		return nil
-	}
-	if m.proc == nil || m.proc.Process == nil {
-		// #3780: the helper is not running, so no snapshot is being
-		// enforced — there is no stale permit to converge. The helper
-		// restart path re-applies the last snapshot. Converged.
-		return nil
-	}
-
-	if err := m.ensureRequiredSnapshotProtocolLocked(cfg); err != nil {
-		if disarmErr := m.disarmSnapshotProtocolFailureLocked(err); disarmErr != nil {
-			slog.Warn("userspace: failed to disarm helper after refusing snapshot publish",
-				"protocol_err", err, "err", disarmErr)
-		}
-		slog.Warn("userspace: refusing snapshot publish to incompatible helper", "err", err)
-		// #3780: the intended new inactive-bit view was NOT applied.
-		// Report failure so the daemon retries on the next scheduler
-		// tick and surfaces the stale-enforcement metric.
-		return fmt.Errorf("userspace: refusing snapshot publish to incompatible helper: %w", err)
-	}
-	next := *m.lastSnapshot
-	nextGeneration := m.generation + 1
-	next.Generation = nextGeneration
-	next.FIBGeneration = m.readFIBGeneration()
-	next.GeneratedAt = time.Now().UTC()
-	next.Config = cfg
-	// #2049: thread the cached dynamic-address feed overlay through the
-	// scheduler-only republish so a CoS scheduler-state change does not
-	// drop feed enforcement until the next full apply. m.mu is already
-	// held here, so read m.feedOverlay directly via cloneFeedOverlay
-	// rather than feedSnapshotOverlay() (which re-locks m.mu and would
-	// deadlock). Matches how the full snapshot build reads the overlay.
-	feedOverlay := cloneFeedOverlay(m.feedOverlay)
-	// #2514: an unresolvable address-book content-ID collision must not
-	// panic the daemon. Abort the scheduler-only republish and retain the
-	// last published snapshot (fail-closed) — the next full apply will
-	// surface the same error to the operator at commit time.
-	policies, err := buildPolicySnapshotsWithSchedulerStateAndFeeds(cfg, activeCopy, feedOverlay)
-	if err != nil {
-		slog.Warn("userspace: skipping policy-scheduler republish; retaining prior snapshot", "err", err)
-		// #3780: the prior snapshot is retained, which for a CLOSING
-		// window means the old permit stays live. Report failure so the
-		// transition is retried until the rebuild succeeds and converges.
-		return fmt.Errorf("userspace: policy snapshot rebuild for scheduler republish: %w", err)
-	}
-	next.Policies = policies
-	// #3261: the policies were rebuilt, so recompute the (feed-aware)
-	// content-rejection diagnostic from the new rules' sentinels; the copied
-	// lastSnapshot value would be stale. Class (ii) (ForwardingSupported /
-	// UnsupportedReasons) is unchanged by a scheduler-only republish.
-	next.Capabilities.PolicyContentRejected = collectPolicyContentRejections(policies)
-	// #1606: refresh the address-book table alongside the policies
-	// so book IDs cited in the new policies always resolve on the
-	// dataplane side.
-	books, _, err := buildAddressBookTableWithFeeds(cfg, feedOverlay)
-	if err != nil {
-		slog.Warn("userspace: skipping policy-scheduler republish; retaining prior snapshot", "err", err)
-		// #3780: same fail-safe as the policy rebuild above — retry.
-		return fmt.Errorf("userspace: address-book rebuild for scheduler republish: %w", err)
-	}
-	next.AddressBooks = books
-
-	publishSnap := next
-	publishSnap.Neighbors = filterPublishableNeighbors(next.Neighbors)
-	var status ProcessStatus
-	// #2124: disarm before publishing an unsupported-config snapshot (see
-	// disarmBeforeUnsupportedPublishLocked). cfg is this snapshot's config.
-	if err := m.disarmBeforeUnsupportedPublishLocked(&publishSnap); err != nil {
-		slog.Warn("userspace: failed to disarm before unsupported-config policy scheduler publish", "err", err)
-		// #3780: disarm failed and the new snapshot was not published —
-		// report failure so the transition retries.
-		return fmt.Errorf("userspace: disarm before unsupported-config policy scheduler publish: %w", err)
-	}
-	if err := m.requestLocked(ControlRequest{Type: "apply_snapshot", Snapshot: &publishSnap}, &status); err != nil {
-		slog.Warn("userspace: failed to publish policy scheduler state", "err", err)
-		// #3780: THE fail-open path from the issue. apply_snapshot did
-		// not land, so the helper keeps the OLD inactive bits — a permit
-		// past its window stays live. Report failure so the daemon
-		// retries autonomously on the next scheduler tick.
-		return fmt.Errorf("userspace: publish policy scheduler snapshot: %w", err)
-	}
-	m.logWgEndpointSetTransitionLocked(&publishSnap, "policy-scheduler")
-	m.generation = nextGeneration
-	m.lastSnapshot = &next
-	m.rebuildNeighborIndex()
-	m.rebuildMonitoredIfindexes()
-	m.publishedSnapshot = next.Generation
-	m.publishedPlanKey = snapshotBindingPlanKey(&next)
-	// #2079: full apply_snapshot succeeded — record the applied snapshot.
-	m.markAppliedSnapshotLocked()
-	if h, ok := snapshotContentHash(&next); ok {
-		m.lastSnapshotHash = h
-	}
-	if err := m.applyHelperStatusLocked(&status); err != nil {
-		// #3780: the snapshot DID land (generation bumped, lastSnapshot
-		// updated above) — the schedule transition converged. A status
-		// re-sync failure is observability only; do NOT force a retry
-		// that would churn an identical snapshot.
-		slog.Warn("userspace: failed to sync helper status after policy scheduler publish", "err", err)
-	}
-	return nil
-}
-
-// routeOverlaySnapshot returns a copy of the cached ip-monitoring
-// route overlay.
-func (m *Manager) routeOverlaySnapshot() []config.RouteOverlayEntry {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	return cloneRouteOverlay(m.routeOverlay)
-}
-
-func cloneRouteOverlay(overlay []config.RouteOverlayEntry) []config.RouteOverlayEntry {
-	if overlay == nil {
-		return nil
-	}
-	out := make([]config.RouteOverlayEntry, len(overlay))
-	copy(out, overlay)
-	return out
-}
-
-// SetRouteOverlay caches the ip-monitoring effective-route overlay for
-// the next full snapshot build WITHOUT publishing. The daemon calls
-// this at the top of applyConfigLocked (holding applySem) so an
-// operator commit while a policy is FAILED rebuilds routes with the
-// active overlay instead of wiping the injected route (#1827, AGY
-// r2-2).
-func (m *Manager) SetRouteOverlay(overlay []config.RouteOverlayEntry) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	m.routeOverlay = cloneRouteOverlay(overlay)
-}
-
-// feedSnapshotOverlay returns a deep copy of the cached dynamic-address
-// feed-prefix overlay (#2049). Read under m.mu, mirroring
-// routeOverlaySnapshot, so the full snapshot build sees a stable view.
-func (m *Manager) feedSnapshotOverlay() map[string][]string {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	return cloneFeedOverlay(m.feedOverlay)
-}
-
-func cloneFeedOverlay(overlay map[string][]string) map[string][]string {
-	if overlay == nil {
-		return nil
-	}
-	out := make(map[string][]string, len(overlay))
-	for name, prefixes := range overlay {
-		cp := make([]string, len(prefixes))
-		copy(cp, prefixes)
-		out[name] = cp
-	}
-	return out
-}
-
-// SetFeedSnapshots caches the dynamic-address feed-prefix overlay for the
-// next full snapshot build WITHOUT publishing (#2049). The daemon calls this
-// at the top of applyConfigLocked (holding applySem) with the live feed
-// snapshots joined to the address-name bindings, so an operator commit OR a
-// feed onUpdate rebuilds the address book with the current feed prefixes.
-// Mirrors SetRouteOverlay. The overlay is deep-copied so the caller may reuse
-// or mutate its map afterwards.
-func (m *Manager) SetFeedSnapshots(overlay map[string][]string) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	m.feedOverlay = cloneFeedOverlay(overlay)
-}
-
-// PublishRouteOverlaySnapshot republishes the userspace snapshot with
-// the routes section rebuilt under the given ip-monitoring overlay
-// (#1827 PR-1b §4.3, modeled on the policy-scheduler partial republish
-// above). It must NOT call Compile (which detaches links / restarts
-// the helper): it clones lastSnapshot, refreshes Routes via
-// buildRouteSnapshots with the overlay, bumps the generation, and
-// reuses the apply_snapshot hash/publish bookkeeping. An overlay whose
-// snapshot content is identical to the last published one is skipped
-// (duplicate-publish skip) and reported as success.
-//
-// Ordering contract (AGY r2-1): the caller (the daemon's routes-only
-// actuator) must call BumpFIBGeneration ONLY after this returns
-// published=true with a nil error — bumping before the helper has the
-// new routes would re-resolve flows against the OLD routes and the
-// later snapshot would not re-invalidate them; bumping after a
-// duplicate-skip would churn established-flow route caches for
-// nothing (Codex PR #1843 MED).
-//
-// schedulerState refreshes the policy snapshots in the same publish
-// when non-nil; nil keeps the manager's current scheduler view.
-func (m *Manager) PublishRouteOverlaySnapshot(cfg *config.Config, overlay []config.RouteOverlayEntry, schedulerState map[string]bool) (published bool, err error) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	// #3760: do NOT advance the cached desired overlay before the publish
-	// is known not to have failed. buildRouteSnapshots below builds
-	// against this local copy; m.routeOverlay is committed only on a
-	// non-error return (deferred commit). A failed apply_snapshot
-	// therefore leaves m.routeOverlay at the last-applied baseline, so
-	// the next actuator sweep rebuilds the same new routes, sees a hash
-	// mismatch against the still-old lastSnapshotHash, and re-publishes
-	// (#3757 dirty-retry contract). Mirrors the mutate-after-success
-	// pattern (#3766/#3742/#3757): the cache never records an overlay the
-	// dataplane never accepted. The nil-error early returns below
-	// (no published snapshot yet, helper not running) still commit the
-	// overlay so the next full apply carries it; the duplicate-skip
-	// return commits a content-equivalent overlay.
-	desiredOverlay := cloneRouteOverlay(overlay)
-	defer func() {
-		if err == nil {
-			m.routeOverlay = desiredOverlay
-		}
-	}()
-
-	if schedulerState != nil {
-		m.policySchedulerActive = copyPolicySchedulerActiveState(schedulerState)
-	}
-
-	if cfg == nil {
-		if m.lastSnapshot == nil {
-			return false, nil
-		}
-		cfg = m.lastSnapshot.Config
-	}
-	if cfg == nil || m.lastSnapshot == nil {
-		// No published snapshot yet: the overlay is cached (deferred
-		// commit) and the next full apply will carry it.
-		return false, nil
-	}
-	if m.proc == nil || m.proc.Process == nil {
-		return false, nil
-	}
-
-	if err := m.ensureRequiredSnapshotProtocolLocked(cfg); err != nil {
-		if disarmErr := m.disarmSnapshotProtocolFailureLocked(err); disarmErr != nil {
-			slog.Warn("userspace: failed to disarm helper after refusing overlay publish",
-				"protocol_err", err, "err", disarmErr)
-		}
-		return false, fmt.Errorf("refusing route overlay publish to incompatible helper: %w", err)
-	}
-
-	next := *m.lastSnapshot
-	nextGeneration := m.generation + 1
-	next.Generation = nextGeneration
-	next.FIBGeneration = m.readFIBGeneration()
-	next.GeneratedAt = time.Now().UTC()
-	next.Config = cfg
-	// #3772 (M9): a transient ip-rule enumeration failure aborts the
-	// overlay publish (fail-closed). The deferred commit above leaves
-	// m.routeOverlay at the last-applied baseline on a non-nil err, so the
-	// next actuator sweep rebuilds and re-publishes (#3757 dirty-retry).
-	next.Routes, err = buildRouteSnapshots(cfg, next.Interfaces, desiredOverlay)
-	if err != nil {
-		return false, fmt.Errorf("build route overlay snapshot: %w", err)
-	}
-
-	// Duplicate-publish skip: identical content (e.g. the actuator ran
-	// twice for the same overlay) does not need a control-socket
-	// round-trip. The content hash excludes Generation/FIBGeneration.
-	if h, ok := snapshotContentHash(&next); ok && h == m.lastSnapshotHash {
-		slog.Debug("userspace: route overlay publish skipped (content unchanged)")
-		return false, nil
-	}
-
-	publishSnap := next
-	publishSnap.Neighbors = filterPublishableNeighbors(next.Neighbors)
-	var status ProcessStatus
-	// #2124: disarm before publishing an unsupported-config snapshot.
-	if err := m.disarmBeforeUnsupportedPublishLocked(&next); err != nil {
-		return false, err
-	}
-	if err := m.requestLocked(ControlRequest{Type: "apply_snapshot", Snapshot: &publishSnap}, &status); err != nil {
-		return false, fmt.Errorf("publish route overlay snapshot: %w", err)
-	}
-	m.logWgEndpointSetTransitionLocked(&publishSnap, "route-overlay")
-	m.generation = nextGeneration
-	m.lastSnapshot = &next
-	m.rebuildNeighborIndex()
-	m.rebuildMonitoredIfindexes()
-	m.publishedSnapshot = next.Generation
-	m.publishedPlanKey = snapshotBindingPlanKey(&next)
-	// #2079: full apply_snapshot succeeded — record the applied snapshot.
-	m.markAppliedSnapshotLocked()
-	if h, ok := snapshotContentHash(&next); ok {
-		m.lastSnapshotHash = h
-	}
-	if err := m.applyHelperStatusLocked(&status); err != nil {
-		slog.Warn("userspace: failed to sync helper status after route overlay publish", "err", err)
-	}
-	slog.Info("userspace: route overlay snapshot published",
-		"generation", next.Generation, "overlay_routes", len(desiredOverlay))
-	return true, nil
-}
-
-func (m *Manager) syncInterfaceAttachments(result *dataplane.CompileResult, snapshot *ConfigSnapshot) {
-	if result == nil {
-		return
-	}
-	allowed := make(map[int]bool)
-	for _, ifindex := range buildUserspaceIngressIfindexes(snapshot) {
-		allowed[int(ifindex)] = true
-	}
-	for ifindex := range m.bpfShim.XDPLinks() {
-		if allowed[ifindex] {
-			continue
-		}
-		if err := m.bpfShim.DetachXDP(ifindex); err != nil {
-			slog.Warn("userspace: detach XDP from non-data interface failed", "ifindex", ifindex, "err", err)
-		}
-	}
-	for ifindex := range m.bpfShim.TCLinks() {
-		if allowed[ifindex] {
-			continue
-		}
-		if err := m.bpfShim.DetachTC(ifindex); err != nil {
-			slog.Warn("userspace: detach TC from non-data interface failed", "ifindex", ifindex, "err", err)
-		}
-	}
-}
-
-func (m *Manager) readFIBGeneration() uint32 {
-	fibGenMap := m.bpfShim.Map("fib_gen_map")
-	if fibGenMap == nil {
-		return 0
-	}
-	var (
-		key uint32
-		gen uint32
-	)
-	if err := fibGenMap.Lookup(key, &gen); err != nil {
-		return 0
-	}
-	return gen
-}
-
-func configHasScheduledPolicy(cfg *config.Config) bool {
-	if cfg == nil {
-		return false
-	}
-	for _, zpp := range cfg.Security.Policies {
-		if zpp == nil {
-			continue
-		}
-		for _, pol := range zpp.Policies {
-			if pol != nil && pol.SchedulerName != "" {
-				return true
-			}
-		}
-	}
-	for _, pol := range cfg.Security.GlobalPolicies {
-		if pol != nil && pol.SchedulerName != "" {
-			return true
-		}
-	}
-	return false
-}
-
-func (m *Manager) ensurePolicySchedulerProtocolLocked(cfg *config.Config) error {
-	if !configHasScheduledPolicy(cfg) {
-		return nil
-	}
-	if m.lastStatus.ConfigSnapshotProtocolVersion >= ProtocolVersion {
-		return nil
-	}
-	var status ProcessStatus
-	if err := m.requestLocked(ControlRequest{Type: "status"}, &status); err == nil {
-		m.recordHelperStatusLocked(&status)
-		if status.ConfigSnapshotProtocolVersion >= ProtocolVersion {
-			return nil
-		}
-	}
-	return fmt.Errorf(
-		"%w: helper config snapshot protocol version %d < required %d for policy scheduler snapshots",
-		ErrPolicySchedulerProtocolIncompatible,
-		m.lastStatus.ConfigSnapshotProtocolVersion,
-		ProtocolVersion,
-	)
-}
-
-func (m *Manager) ensurePersistentSourceNATProtocolLocked(cfg *config.Config) error {
-	if !userspaceConfigUsesPersistentSourceNAT(cfg) {
-		return nil
-	}
-	if m.lastStatus.ConfigSnapshotProtocolVersion >= ProtocolVersion {
-		return nil
-	}
-	var status ProcessStatus
-	if err := m.requestLocked(ControlRequest{Type: "status"}, &status); err == nil {
-		m.recordHelperStatusLocked(&status)
-		if status.ConfigSnapshotProtocolVersion >= ProtocolVersion {
-			return nil
-		}
-	}
-	return fmt.Errorf(
-		"%w: helper config snapshot protocol version %d < required %d for persistent source NAT snapshots",
-		ErrPersistentSourceNATProtocolIncompatible,
-		m.lastStatus.ConfigSnapshotProtocolVersion,
-		ProtocolVersion,
-	)
-}
-
-func (m *Manager) ensureRequiredSnapshotProtocolLocked(cfg *config.Config) error {
-	if err := m.ensurePolicySchedulerProtocolLocked(cfg); err != nil {
-		return err
-	}
-	return m.ensurePersistentSourceNATProtocolLocked(cfg)
-}
-
-func (m *Manager) recordHelperStatusLocked(status *ProcessStatus) {
-	status.DataplaneMode = m.mode.String()
-	status.ConfiguredMode = m.configuredMode.String()
-	status.EntryPrograms = m.entryProgramsLocked()
-	status.DegradedPathCounters = m.readDegradedPathStatsLocked()
-	// #3261: stamp the manager-owned snapshot-reject diagnostic onto the status
-	// the helper round-trip cannot carry (the helper has no PolicyContentRejected
-	// field and strips it on decode). This keeps the rejected-snapshot state
-	// observable in `show`/metrics even though the helper reports the
-	// previous-good capabilities.
-	status.LastSnapshotRejectReasons = append([]string(nil), m.lastSnapshotRejectReasons...)
-	// #3719: stamp the manager-owned zone-id-collision diagnostic (the helper
-	// cannot carry it — the colliding zone was dropped before the wire) so the
-	// degraded zone-isolation state is observable in `show`/metrics.
-	status.ZoneIDCollisions = append([]string(nil), m.lastZoneIDCollisions...)
-	if m.eventStream != nil {
-		es := m.eventStream.Status()
-		status.EventStream = &es
-	}
-	m.lastStatus = *status
-}
-
-func (m *Manager) disarmSnapshotProtocolFailureLocked(protocolErr error) error {
-	if m.proc == nil || m.proc.Process == nil {
-		return nil
-	}
-	req := ControlRequest{
-		Type: "set_forwarding_state",
-		Forwarding: &ForwardingControlRequest{
-			Armed: false,
-		},
-	}
-	var status ProcessStatus
-	if err := m.requestLocked(req, &status); err != nil {
-		return fmt.Errorf("userspace: disarm helper after snapshot protocol error: %w", err)
-	}
-	if err := m.applyHelperStatusLocked(&status); err != nil {
-		m.recordHelperStatusLocked(&status)
-		return fmt.Errorf("userspace: sync helper status after snapshot protocol fail-closed disarm: %w", err)
-	}
-	slog.Warn("userspace: disarmed helper after snapshot protocol error", "err", protocolErr)
-	return nil
-}
-
-// bpfKtimeNs returns the current CLOCK_BOOTTIME in nanoseconds, matching
-// the clock used by BPF's bpf_ktime_get_ns() for session Created timestamps.
-func (m *Manager) bpfKtimeNs() uint64 {
-	var ts unix.Timespec
-	_ = unix.ClockGettime(unix.CLOCK_BOOTTIME, &ts)
-	return uint64(ts.Sec)*1_000_000_000 + uint64(ts.Nsec)
-}
-
-func (m *Manager) bumpGeneration() uint64 {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	m.generation++
-	return m.generation
-}
-
-// BumpFIBGeneration updates the BPF FIB generation counter and sends a
-// lightweight FIB generation bump to the userspace helper. If kernel neighbors
-// changed since the last publish, an incremental neighbor update is sent first.
-// This avoids the full buildSnapshot() + apply_snapshot round-trip that was the
-// primary source of control socket contention during route convergence.
-//
-// Error contract (#1844): a non-nil error means the helper-side
-// invalidation is NOT confirmed (shim map bump or the
-// bump_fib_generation control message failed) — callers with retry
-// semantics (the ip-monitoring actuator's pendingFIBBump) must retry.
-// The helperless / no-snapshot early returns are SUCCESS (nil): with
-// no published snapshot there are no cached flow routes to invalidate,
-// and the next full apply carries its own invalidation. A failed
-// incremental NEIGHBOR update is deliberately NOT an error — it has
-// its own retry semantics (the cached neighbor view is only advanced
-// on success, so the next bump re-diffs and re-sends).
-func (m *Manager) BumpFIBGeneration() (uint32, error) {
-	newGen, shimErr := m.bpfShim.BumpFIBGeneration()
-
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	if m.lastSnapshot == nil || m.lastSnapshot.Config == nil {
-		return newGen, nil
-	}
-	if m.proc == nil || m.proc.Process == nil {
-		return newGen, nil
-	}
-
-	// Update the cached snapshot's FIB generation without rebuilding.
-	m.lastSnapshot.FIBGeneration = newGen
-	m.generation++
-	m.lastSnapshot.Generation = m.generation
-
-	// #1197 v4 (Codex code-review v3 #1): refresh the monitored
-	// ifindex cache UNCONDITIONALLY — link recreation can happen
-	// without any neighbor diff (operator unplugs cable, kernel
-	// rebinds a VLAN, etc.), and a stale cache silently drops
-	// events on the new ifindex until the next config commit.
-	m.rebuildMonitoredIfindexes()
-
-	// Check if kernel neighbors changed — if so, push an incremental update.
-	// #1197: use forwarding-effective diff so REACHABLE↔STALE aging churn
-	// doesn't trigger unnecessary publishes; filter publish payload to
-	// publishable-only entries (matches userspace-dp accept rules).
-	newNeighbors := buildNeighborSnapshots(m.lastSnapshot.Config)
-	if !neighborsEqualForwarding(m.lastSnapshot.Neighbors, newNeighbors) {
-		publishable := filterPublishableNeighbors(newNeighbors)
-		var status ProcessStatus
-		if err := m.requestLocked(ControlRequest{
-			Type:            "update_neighbors",
-			Neighbors:       publishable,
-			NeighborReplace: true,
-		}, &status); err != nil {
-			slog.Warn("userspace: failed to publish neighbor update", "err", err)
-		} else {
-			// Only update cached neighbors after successful publish so
-			// a transient failure doesn't suppress future retries.
-			m.lastSnapshot.Neighbors = newNeighbors
-			m.rebuildNeighborIndex() // #1197
-		}
-	}
-
-	// Send lightweight FIB generation bump — no full snapshot rebuild.
-	var status ProcessStatus
-	if err := m.requestLocked(ControlRequest{
-		Type: "bump_fib_generation",
-		Snapshot: &ConfigSnapshot{
-			// #3767 H4: the helper now version-gates bump_fib_generation
-			// exactly like apply_snapshot. Stamp the protocol version so a
-			// legitimate route-only bump is accepted; an unversioned (0)
-			// message is rejected as a mixed-version / corrupt client.
-			Version:       ProtocolVersion,
-			FIBGeneration: newGen,
-		},
-	}, &status); err != nil {
-		slog.Warn("userspace: failed to bump FIB generation", "err", err)
-		return newGen, fmt.Errorf("bump fib generation: %w", err)
-	}
-	return newGen, shimErr
-}
-
-// neighborIndexKey is the (ifindex, ip-string) key for the
-// O(1) neighbor lookup index used by the daemon's listener hot
-// path. ip-string is used (not net.IP) so map equality is well-
-// defined for both v4 and v6 representations.
-type neighborIndexKey struct {
-	ifindex int
-	ip      string
-}
-
-// filterPublishableNeighbors returns only the entries
-// userspace-dp will accept (per neighborSnapshotPublishable).
-func filterPublishableNeighbors(neighbors []NeighborSnapshot) []NeighborSnapshot {
-	out := make([]NeighborSnapshot, 0, len(neighbors))
-	for _, n := range neighbors {
-		if neighborSnapshotPublishable(n) {
-			out = append(out, n)
-		}
-	}
-	return out
-}
-
-// rebuildNeighborIndex updates m.neighborIndex from the current
-// m.lastSnapshot.Neighbors slice. Caller MUST hold m.mu. Called
-// after every assignment to lastSnapshot.Neighbors.
-//
-// Codex code-review v2 #1: index ONLY publishable entries.
-// Indexing raw entries causes a bug: a raw failed→reachable
-// transition on the same MAC would return existing.MAC == new.MAC
-// from LookupSnapshotNeighbor → shouldTriggerRegen returns false
-// → snapshot stays out of date until 60s safety tick. By indexing
-// only publishable entries, an unpublishable→publishable
-// transition presents as "no existing entry" → trigger fires.
-func (m *Manager) rebuildNeighborIndex() {
-	if m.lastSnapshot == nil {
-		m.neighborIndex = nil
-		return
-	}
-	idx := make(map[neighborIndexKey]*NeighborSnapshot,
-		len(m.lastSnapshot.Neighbors))
-	for i := range m.lastSnapshot.Neighbors {
-		n := &m.lastSnapshot.Neighbors[i]
-		if !neighborSnapshotPublishable(*n) {
-			continue
-		}
-		idx[neighborIndexKey{n.Ifindex, n.IP}] = n
-	}
-	m.neighborIndex = idx
-}
-
-// RegenerateNeighborSnapshot rebuilds the in-memory neighbor
-// snapshot from current kernel ARP/NDP state and publishes any
-// forwarding-relevant changes to userspace-dp.
-//
-// #1197: this is the event-driven entry point used by the
-// daemon's RTM_NEWNEIGH/DELNEIGH listener (and the 60s safety
-// reconciliation tick) to keep the userspace-dp neighbor table
-// in sync with the kernel without depending on the buggy
-// preinstall mechanism.
-//
-// Forwarding-effective diff (key, MAC, publishable-bit) decides
-// whether to publish; raw NUD state (REACHABLE↔STALE) is ignored
-// to avoid republish churn.
-func (m *Manager) RegenerateNeighborSnapshot() {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	if m.lastSnapshot == nil || m.lastSnapshot.Config == nil {
-		return
-	}
-	if m.proc == nil || m.proc.Process == nil {
-		return
-	}
-	// #1197 v4 (Codex code-review v3 #1): refresh monitored
-	// ifindex cache unconditionally — a regen call may be
-	// triggered by the safety tick precisely because a link
-	// changed, and the listener filter must reflect that even
-	// if neighbor entries didn't diff.
-	m.rebuildMonitoredIfindexes()
-
-	newNeighbors := buildNeighborSnapshots(m.lastSnapshot.Config)
-	if neighborsEqualForwarding(m.lastSnapshot.Neighbors, newNeighbors) {
-		return
-	}
-	publishable := filterPublishableNeighbors(newNeighbors)
-	var status ProcessStatus
-	if err := m.requestLocked(ControlRequest{
-		Type:            "update_neighbors",
-		Neighbors:       publishable,
-		NeighborReplace: true,
-	}, &status); err != nil {
-		slog.Warn("userspace: failed to publish neighbor regeneration", "err", err)
-		return
-	}
-	m.lastSnapshot.Neighbors = newNeighbors
-	m.rebuildNeighborIndex() // #1197 (after publish success)
-	m.generation++
-	m.lastSnapshot.Generation = m.generation
-	// Copilot review: advance publishedSnapshot + refresh
-	// lastSnapshotHash. Otherwise the status loop sees the
-	// bumped generation as unpublished and may force a redundant
-	// apply_snapshot, AND any churn in filtered-out rows could
-	// leak through hash-dedup.
-	m.publishedSnapshot = m.lastSnapshot.Generation
-	if h, ok := snapshotContentHash(m.lastSnapshot); ok {
-		m.lastSnapshotHash = h
-	}
-}
-
-// LookupSnapshotNeighbor returns a copy of the snapshot's
-// current entry for (ifindex, ip), or nil if not present. The
-// returned snapshot is a value copy — safe to read after the
-// lock is released, and avoids the (currently no-op) mutation
-// hazard of returning an internal pointer.
-//
-// Codex code-review v2 #4: previous version returned a defensive
-// pointer via heap copy. Caller (shouldTriggerRegen) only reads
-// the MAC immediately while still under m.mu, so a pointer is
-// safe. But the API surface is cleaner as a value (caller can
-// hold it across other lock-acquiring calls without aliasing
-// concerns), so we keep the value-copy semantics — just skip
-// the heap-allocated *NeighborSnapshot wrapping.
-//
-// Index covers ONLY publishable entries (#1 v2 fix), so a hit
-// here means userspace-dp has been told about this entry.
-func (m *Manager) LookupSnapshotNeighbor(ifindex int, ip net.IP) *NeighborSnapshot {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	if m.neighborIndex == nil {
-		return nil
-	}
-	entry, ok := m.neighborIndex[neighborIndexKey{ifindex, ip.String()}]
-	if !ok {
-		return nil
-	}
-	out := *entry
-	return &out
-}
-
-// IsMonitoredIfindex returns true if the given link index
-// belongs to a configured interface that buildNeighborSnapshots
-// would iterate. O(1) hash-map lookup under m.mu.
-//
-// Codex code-review v2 #2: previous version returned a copy of
-// the whole map, which made the listener hot path O(configured-
-// interfaces) plus heap churn per event. This direct lookup
-// avoids both.
-func (m *Manager) IsMonitoredIfindex(ifindex int) bool {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	if m.monitoredIfindexes == nil {
-		return false
-	}
-	_, ok := m.monitoredIfindexes[ifindex]
-	return ok
-}
-
-// rebuildMonitoredIfindexes updates m.monitoredIfindexes from
-// the active config. Caller MUST hold m.mu.
-//
-// Codex code-review v2 #3: previous version was called only on
-// full snapshot assignment. Neighbor-only updates (BumpFIBGeneration,
-// RegenerateNeighborSnapshot) didn't refresh the cache, so a
-// configured link recreated under a new ifindex could have its
-// events dropped. Now called from every neighbor-related update
-// path that may reflect a new ifindex.
-func (m *Manager) rebuildMonitoredIfindexes() {
-	if m.lastSnapshot == nil || m.lastSnapshot.Config == nil {
-		m.monitoredIfindexes = nil
-		return
-	}
-	m.monitoredIfindexes = MonitoredInterfaceLinkIndexes(m.lastSnapshot.Config)
-}
-
-// ForEachSnapshotNeighbor invokes fn for every PUBLISHABLE
-// neighbor entry in the current snapshot (i.e., entries
-// userspace-dp accepted into its forwarding table).
-//
-// #1197 (Copilot review): the existing SnapshotNeighbors() walks
-// raw lastSnapshot.Neighbors which can include filtered-out
-// entries (FAILED/INCOMPLETE/none). For force-probe target
-// collection we want only the entries the dataplane is actually
-// using, so we walk neighborIndex (publishable-only).
-func (m *Manager) ForEachSnapshotNeighbor(fn func(ifindex int, ip net.IP)) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	for k, n := range m.neighborIndex {
-		ip := net.ParseIP(n.IP)
-		if ip == nil {
-			continue
-		}
-		fn(k.ifindex, ip)
-	}
-}
-
-// SnapshotHasIfindex returns true if the current snapshot
-// contains any neighbor entry on the given ifindex. Used by the
-// daemon's listener filter as a fallback for runtime ifindex
-// drift. O(N) scan over the neighborIndex but the listener
-// already pays O(1) for the LookupSnapshotNeighbor; this fallback
-// only fires when the config-derived monitored set doesn't
-// contain the ifindex (rare, e.g., link disappeared).
-func (m *Manager) SnapshotHasIfindex(ifindex int) bool {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	for k := range m.neighborIndex {
-		if k.ifindex == ifindex {
-			return true
-		}
-	}
-	return false
-}
-
-// SnapshotNeighbors returns the neighbor entries from the last published
-// snapshot. Used by the daemon to pre-install kernel ARP entries on RG
-// activation so failback doesn't drop packets during ARP resolution.
-func (m *Manager) SnapshotNeighbors() []struct {
-	Ifindex int
-	IP      net.IP
-	MAC     net.HardwareAddr
-	Family  int
-} {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	if m.lastSnapshot == nil {
-		return nil
-	}
-	var result []struct {
-		Ifindex int
-		IP      net.IP
-		MAC     net.HardwareAddr
-		Family  int
-	}
-	for _, n := range m.lastSnapshot.Neighbors {
-		if n.Ifindex <= 0 || n.MAC == "" || n.IP == "" {
-			continue
-		}
-		mac, err := net.ParseMAC(n.MAC)
-		if err != nil {
-			continue
-		}
-		ip := net.ParseIP(n.IP)
-		if ip == nil {
-			continue
-		}
-		family := netlink.FAMILY_V4
-		if ip.To4() == nil {
-			family = netlink.FAMILY_V6
-		}
-		result = append(result, struct {
-			Ifindex int
-			IP      net.IP
-			MAC     net.HardwareAddr
-			Family  int
-		}{
-			Ifindex: n.Ifindex,
-			IP:      ip,
-			MAC:     mac,
-			Family:  family,
-		})
-	}
-	return result
-}
-
-func (m *Manager) Status() (ProcessStatus, error) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	if m.proc == nil {
-		if m.lastStatus.PID != 0 {
-			return m.lastStatus, nil
-		}
-		return ProcessStatus{}, errors.New("userspace dataplane helper not running")
-	}
-
-	var status ProcessStatus
-	if err := m.requestLocked(ControlRequest{Type: "status"}, &status); err != nil {
-		if m.lastStatus.PID != 0 {
-			return m.lastStatus, err
-		}
-		return ProcessStatus{}, err
-	}
-	if err := m.applyHelperStatusLocked(&status); err != nil {
-		return status, err
-	}
-	return status, nil
-}
-
-// CachedStatus returns the most recent ProcessStatus captured by a
-// control-socket round-trip (primarily the 1 Hz statusLoop, which
-// refreshes m.lastStatus every second via applyHelperStatusLocked),
-// WITHOUT issuing a new control-socket request. The bool is false
-// when no status has been captured yet (helper not started or never
-// polled), in which case the caller should hold its previous values.
-//
-// This exists so low-priority periodic consumers -- specifically the
-// fwdstatus CPU sampler -- can read worker telemetry off the shared
-// status poll instead of issuing their own redundant 1 Hz "status"
-// request, which would double the control-socket status rate and
-// starve session installs during bulk sync (#3970; CLAUDE.md
-// "Control socket contention"). Unlike Status(), this MUST NOT touch
-// the control socket.
-func (m *Manager) CachedStatus() (ProcessStatus, bool) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	if m.lastStatus.PID == 0 {
-		return ProcessStatus{}, false
-	}
-	return m.lastStatus, true
-}
-
-func (m *Manager) SetForwardingArmed(armed bool) (ProcessStatus, error) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	if m.proc == nil {
-		return ProcessStatus{}, errors.New("userspace dataplane helper not running")
-	}
-	if armed && !m.lastStatus.Capabilities.ForwardingSupported {
-		if len(m.lastStatus.Capabilities.UnsupportedReasons) == 0 {
-			return m.lastStatus, errors.New("userspace live forwarding is not supported for the current configuration")
-		}
-		return m.lastStatus, fmt.Errorf(
-			"userspace live forwarding is not supported: %s",
-			strings.Join(m.lastStatus.Capabilities.UnsupportedReasons, "; "),
-		)
-	}
-	var status ProcessStatus
-	req := ControlRequest{
-		Type: "set_forwarding_state",
-		Forwarding: &ForwardingControlRequest{
-			Armed: armed,
-		},
-	}
-	if err := m.requestLocked(req, &status); err != nil {
-		return ProcessStatus{}, err
-	}
-	if err := m.applyHelperStatusLocked(&status); err != nil {
-		return status, err
-	}
-	return status, nil
-}
-
-func (m *Manager) SetQueueState(queueID uint32, registered, armed bool) (ProcessStatus, error) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	if m.proc == nil {
-		return ProcessStatus{}, errors.New("userspace dataplane helper not running")
-	}
-	var status ProcessStatus
-	req := ControlRequest{
-		Type: "set_queue_state",
-		Queue: &QueueControlRequest{
-			QueueID:    queueID,
-			Registered: registered,
-			Armed:      armed,
-		},
-	}
-	if err := m.requestLocked(req, &status); err != nil {
-		return ProcessStatus{}, err
-	}
-	if err := m.applyHelperStatusLocked(&status); err != nil {
-		return status, err
-	}
-	return status, nil
-}
-
-func (m *Manager) SetBindingState(slot uint32, registered, armed bool) (ProcessStatus, error) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	if m.proc == nil {
-		return ProcessStatus{}, errors.New("userspace dataplane helper not running")
-	}
-	var status ProcessStatus
-	req := ControlRequest{
-		Type: "set_binding_state",
-		Binding: &BindingControlRequest{
-			Slot:       slot,
-			Registered: registered,
-			Armed:      armed,
-		},
-	}
-	if err := m.requestLocked(req, &status); err != nil {
-		return ProcessStatus{}, err
-	}
-	if err := m.applyHelperStatusLocked(&status); err != nil {
-		return status, err
-	}
-	return status, nil
-}
-
-func (m *Manager) InjectPacket(req InjectPacketRequest) (ProcessStatus, error) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	if m.proc == nil {
-		return ProcessStatus{}, errors.New("userspace dataplane helper not running")
-	}
-	if err := validateInjectPacketRequestForHelper(req, m.lastStatus); err != nil {
-		return ProcessStatus{}, err
-	}
-	var status ProcessStatus
-	if err := m.requestLocked(ControlRequest{Type: "inject_packet", Packet: &req}, &status); err != nil {
-		return ProcessStatus{}, err
-	}
-	if err := m.applyHelperStatusLocked(&status); err != nil {
-		return status, err
-	}
-	return status, nil
-}
-
-func (m *Manager) DrainSessionDeltas(max uint32) ([]SessionDeltaInfo, ProcessStatus, error) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	if m.proc == nil {
-		return nil, ProcessStatus{}, errors.New("userspace dataplane helper not running")
-	}
-	resp, err := m.requestDetailedLocked(ControlRequest{
-		Type: "drain_session_deltas",
-		SessionDeltas: &SessionDeltaDrainRequest{
-			Max: max,
-		},
-	})
-	if err != nil {
-		return nil, ProcessStatus{}, err
-	}
-	var status ProcessStatus
-	if resp.Status != nil {
-		status = *resp.Status
-		if err := m.applyHelperStatusLocked(&status); err != nil {
-			return resp.SessionDeltas, status, err
-		}
-	}
-	return resp.SessionDeltas, status, nil
-}
-
-func (m *Manager) ExportOwnerRGSessions(rgIDs []int, max uint32) ([]SessionDeltaInfo, ProcessStatus, error) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	if m.proc == nil {
-		return nil, ProcessStatus{}, errors.New("userspace dataplane helper not running")
-	}
-	resp, err := m.requestDetailedLocked(ControlRequest{
-		Type: "export_owner_rg_sessions",
-		SessionExport: &SessionExportRequest{
-			OwnerRGs: rgIDs,
-			Max:      max,
-		},
-	})
-	if err != nil {
-		return nil, ProcessStatus{}, err
-	}
-	var status ProcessStatus
-	if resp.Status != nil {
-		status = *resp.Status
-		if err := m.applyHelperStatusLocked(&status); err != nil {
-			return resp.SessionDeltas, status, err
-		}
-	}
-	return resp.SessionDeltas, status, nil
 }
